Extracts Msf::DBManager::Import::Foundstone
MSP-11124
