+++ lib/msf/core/db_manager/import.rb
@@ -12,7 +12,6 @@ require 'uri'
 #
 
 require 'packetfu'
-require 'rex/parser/foundstone_nokogiri'
 require 'rex/parser/fusionvm_nokogiri'
 require 'rex/parser/mbsa_nokogiri'
 require 'rex/parser/nessus_xml'
@@ -33,6 +32,7 @@ module Msf::DBManager::Import
   autoload :Appscan, 'msf/core/db_manager/import/appscan'
   autoload :Burp, 'msf/core/db_manager/import/burp'
   autoload :CI, 'msf/core/db_manager/import/ci'
+  autoload :Foundstone, 'msf/core/db_manager/import/foundstone'
   autoload :IP360, 'msf/core/db_manager/import/ip360'
   autoload :MsfXml, 'msf/core/db_manager/import/msf_xml'
   autoload :Qualys, 'msf/core/db_manager/import/qualys'
@@ -42,6 +42,7 @@ module Msf::DBManager::Import
   include Msf::DBManager::Import::Appscan
   include Msf::DBManager::Import::Burp
   include Msf::DBManager::Import::CI
+  include Msf::DBManager::Import::Foundstone
   include Msf::DBManager::Import::IP360
   include Msf::DBManager::Import::MsfXml
   include Msf::DBManager::Import::Qualys
@@ -338,36 +339,6 @@ module Msf::DBManager::Import
     end
 
     raise DBImportError.new("Could not automatically determine file type")
-  end
-
-  def import_foundstone_noko_stream(args={},&block)
-    if block
-      doc = Rex::Parser::FoundstoneDocument.new(args,framework.db) {|type, data| yield type,data }
-    else
-      doc = Rex::Parser::FoundstoneDocument.new(args,self)
-    end
-    parser = ::Nokogiri::XML::SAX::Parser.new(doc)
-    parser.parse(args[:data])
-  end
-
-  def import_foundstone_xml(args={}, &block)
-    bl = validate_ips(args[:blacklist]) ? args[:blacklist].split : []
-    wspace = args[:wspace] || workspace
-    if Rex::Parser.nokogiri_loaded
-      parser = "Nokogiri v#{::Nokogiri::VERSION}"
-      noko_args = args.dup
-      noko_args[:blacklist] = bl
-      noko_args[:wspace] = wspace
-      if block
-        yield(:parser, parser)
-        import_foundstone_noko_stream(noko_args) {|type, data| yield type,data}
-      else
-        import_foundstone_noko_stream(noko_args)
-      end
-      return true
-    else # Sorry
-      raise DBImportError.new("Could not import due to missing Nokogiri parser. Try 'gem install nokogiri'.")
-    end
   end
 
   def import_fusionvm_xml(args={})
